+++ components/dashboard/types.ts
@@ -1,0 +1,11 @@
+type Sentiment = {
+  magnitude: number
+  score: number
+}
+
+type Text = {
+  content: string
+  beginOffset: number
+}
+
+export type { Sentiment, Text }
